tools/sample-apps/tweet-analyzer/requirements.txt
flask-restful
flask
psycopg2
sqlalchemy
watchdog
eventlet
websocket_client
